requirements.txt
# Core ML / Deep Learning
torch>=2.0
torchvision>=0.15
tensorflow>=2.15

# Image processing
opencv-python>=4.8
Pillow>=10.0

# Data science
numpy>=1.24
pandas>=1.5
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3

# Jupyter
jupyter>=1.0
notebook>=7.0

# Utilities
tqdm>=4.65
pyyaml>=6.0
python-dotenv>=1.0

